Gen 1 TM Learnsets for Gen 2 Mons
Used in order of relevance:
1) Gen 2 TM lists (for identical moves like Blizzard et al)
2) Gen 2 Level-Up/Egg Lists
3) Gen 3 Tutor List (which included most Gen 1 TMs)
4) Gen 4+ TM/Tutor Lists (if no other were available earlier
5) Best judgment

Care was done to NOT give moves to mons that only get a move in Gen 4+ where possible to preserve some of that classic Gen 2 jank. Revisions can be made down the line if anyone thinks anyone really should or shouldn't get something in particular.